feat: support different number of incoming samples and acq samples (#42)
In order to support different number of samples for measurement
and acquisition, the code phase might need to be adjusted to align
the start of tracking with the previous acquisition. The reason
for this is that the samples for acquisition are buffered in a
FIFO buffer. Hence, when the buffer is full and more samples
are pushed into the buffer, the code phase needs to re-aligned
with tracking, which starts with the first incoming sample.